Rename src files to something that isnt so samply
