+++ backend/src/Application.kt
@@ -1,11 +1,10 @@
 package com.conferenceplatform
 
+import common.hello
 import io.ktor.application.*
 import io.ktor.response.*
 import io.ktor.request.*
 import io.ktor.routing.*
-
-import sample.*
 
 fun main(args: Array<String>): Unit = io.ktor.server.netty.EngineMain.main(args)
