+++ common/src/commonMain/kotlin/common/Platform.kt
@@ -1,0 +1,7 @@
+package common
+
+expect object Platform {
+    val name: String
+}
+
+fun hello(): String = "Conference Platform for ${Platform.name}"
