--- common/src/commonMain/kotlin/sample/Sample.kt
@@ -1,11 +1,0 @@
-package sample
-
-expect class Sample() {
-    fun checkMe(): Int
-}
-
-expect object Platform {
-    val name: String
-}
-
-fun hello(): String = "Hello from ${Platform.name}"
