+++ common/src/commonTest/kotlin/common/SampleTests.kt
@@ -1,0 +1,8 @@
+package common
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTests {
+
+}
