+++ common/src/iosMain/kotlin/common/Platform.kt
@@ -1,0 +1,5 @@
+package common
+
+actual object Platform {
+    actual val name: String = "iOS"
+}
