--- common/src/iosMain/kotlin/sample/SampleIOS.kt
@@ -1,9 +1,0 @@
-package sample
-
-actual class Sample {
-    actual fun checkMe() = 7
-}
-
-actual object Platform {
-    actual val name: String = "iOS"
-}
