+++ common/src/iosTest/kotlin/common/PlatformTests.kt
@@ -1,11 +1,11 @@
-package sample
+package common
 
 import kotlin.test.Test
 import kotlin.test.assertTrue
 
-class SampleTestsAndroid {
+class SampleTestsIOS {
     @Test
     fun testHello() {
-        assertTrue("JVM" in hello())
+        assertTrue("iOS" in common.hello())
     }
 }
