--- common/src/jvmMain/kotlin/SampleJVM.kt
@@ -1,9 +1,0 @@
-package sample
-
-actual class Sample {
-    actual fun checkMe() = 42
-}
-
-actual object Platform {
-    actual val name: String = "JVM"
-}
