+++ common/src/jvmTest/kotlin/common/PlatformTests.kt
@@ -1,11 +1,11 @@
-package sample
+package common
 
 import kotlin.test.Test
 import kotlin.test.assertTrue
 
-class SampleTestsJS {
+class SampleTestsAndroid {
     @Test
     fun testHello() {
-        assertTrue("JS" in hello())
+        assertTrue("JVM" in hello())
     }
 }
